requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export -o requirements.txt --no-hashes
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via langchain-community
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via aiohttp
beautifulsoup4==4.13.4
    # via google
boto3==1.38.8
    # via ai-contest-hanare-banare
botocore==1.38.8
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   cryptography
    #   pynacl
    #   zstandard
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   pytest
    #   tqdm
cryptography==44.0.3
    # via pyjwt
dataclasses-json==0.6.7
    # via langchain-community
deprecated==1.2.18
    # via pygithub
distro==1.9.0
    # via openai
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpy==0.0.5
    # via ai-contest-hanare-banare
gitpython==3.1.44
    # via ai-contest-hanare-banare
google==3.0.0
    # via ai-contest-hanare-banare
google-api-core==2.24.2
    # via google-api-python-client
google-api-python-client==2.168.0
    # via ai-contest-hanare-banare
google-auth==2.39.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.70.0
    # via google-api-core
greenlet==3.2.1 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
    # via sqlalchemy
h11==0.16.0
    # via httpcore
html2text==2025.4.15
    # via ai-contest-hanare-banare
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jiter==0.9.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.24
    # via
    #   ai-contest-hanare-banare
    #   langchain-community
langchain-community==0.3.22
    # via ai-contest-hanare-banare
langchain-core==0.3.56
    # via
    #   ai-contest-hanare-banare
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.14
    # via ai-contest-hanare-banare
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.37
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
marshmallow==3.26.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.2.5
    # via
    #   ai-contest-hanare-banare
    #   langchain-community
    #   pandas
openai==1.76.0
    # via langchain-openai
orjson==3.10.16 ; platform_python_implementation != 'PyPy'
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   pytest
pandas==2.2.3
    # via ai-contest-hanare-banare
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via google-api-core
protobuf==6.30.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   ai-contest-hanare-banare
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.9.1
    # via
    #   ai-contest-hanare-banare
    #   langchain-community
pygithub==2.6.1
    # via ai-contest-hanare-banare
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via pygithub
pynacl==1.5.0
    # via pygithub
pyparsing==3.2.3
    # via httplib2
pytest==8.3.5
    # via ai-contest-hanare-banare
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.1.0
    # via pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   google-api-core
    #   langchain
    #   langchain-community
    #   langsmith
    #   pygithub
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==14.0.0
    # via typer
rsa==4.9.1
    # via google-auth
ruff==0.11.8
    # via ai-contest-hanare-banare
s3transfer==0.12.0
    # via boto3
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
slack-sdk==3.35.0
    # via ai-contest-hanare-banare
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
sqlalchemy==2.0.40
    # via
    #   langchain
    #   langchain-community
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.9.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typer==0.15.3
    # via ai-contest-hanare-banare
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   sqlalchemy
    #   typer
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.4.0
    # via
    #   botocore
    #   pygithub
    #   requests
wrapt==1.17.2
    # via deprecated
yarl==1.20.0
    # via aiohttp
zstandard==0.23.0
    # via langsmith
